Stock-Information/requirements.txt
asn1crypto==0.24.0
astroid==2.3.2
attrs==17.4.0
Automat==0.6.0
backports.csv==1.0.7
blinker==1.4
CacheControl==0.12.6
cachetools==3.1.1
certifi==2020.6.20
cffi==1.13.1
chardet==3.0.4
click==6.7
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.6.1
cycler==0.10.0
dialogflow==1.0.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.3
finnhub-python==1.1.4
firebase-admin==4.3.0
flaky==3.6.1
Flask==1.1.2
Flask-Cors==3.0.8
gcp-devrel-py-tools==0.0.15
get-all-tickers==1.6
gevent==20.6.2
google-api-core==1.14.3
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
google-cloud-iot==0.3.0
google-cloud-pubsub==1.0.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.6.0
greenlet==0.4.16
grpc-google-iam-v1==0.12.3
grpcio==1.24.3
httplib2==0.18.0
hyperlink==17.3.1
idna==2.9
incremental==16.10.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.2.0
language-selector==0.1
lazy-object-proxy==1.4.3
lxml==4.5.1
MarkupSafe==1.0
matplotlib==3.2.2
mccabe==0.6.1
mpl-finance==0.10.1
msgpack==1.0.0
multitasking==0.0.9
netifaces==0.10.4
numpy==1.19.0
oauthlib==2.0.6
packaging==19.2
paho-mqtt==1.4.0
PAM==0.4.2
pandas==1.0.5
pandas-datareader==0.8.1
plotly==4.8.1
protobuf==3.12.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.19
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pylint==2.4.3
pyOpenSSL==17.5.0
pyparsing==2.4.2
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytimeparse==1.1.8
pytz==2020.1
pyxdg==0.26
PyYAML==5.1
realtime-stock==1.0.0
requests==2.24.0
requests-unixsocket==0.1.5
retrying==1.3.3
rsa==4.0
SecretStorage==2.3.1
service-identity==16.0.0
six==1.12.0
ssh-import-id==5.7
systemd-python==234
ta==0.5.25
tablib==2.0.0
twelvedata==1.0.1
Twisted==17.9.0
typed-ast==1.4.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.25.9
virtualenv==16.7.7
websocket==0.2.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.11.2
yahoo-earnings-calendar==0.6.0
Yahoo-ticker-downloader==3.0.1
yfinance==0.1.54
zope.event==4.4
zope.interface==5.1.0
